+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog for euler-hs
 
+## [2.1.0.0] - 2021-04-12
+
+* `ManagerSelector` is now a newtype.
+* Add `CertificateRegistrationError` data type.
+* Add 'withSelfSignedFlowRuntime` function for registering self-signed
+  certificates in a `FlowRuntime`.
+
+## [2.0.4.4] - 2021-04-13
+
+* Rename `changeLoggerContext` to `withLoggerContext`
+* Add `updateLoggerContext`
+
+## [2.0.4.3] - 2021-04-12
+
+* Add `changeLoggerContext` to EulerHS.Extra.Language
+
+## [2.0.4.2] - 2021-04-07
+
+* Fixed a bug with MVar blocking when disposing `Async` logger.
+
 ## [2.0.4.1] - 2021-03-03
 
 * Remove waste entries.